+++ LocalPackages/DesignSystem/Tests/DesignSystemTests/PassIconTests.swift
@@ -98,7 +98,6 @@ final class PassIconTests: XCTestCase {
         _ = PassIcon.tabProfileTrialSelected
         _ = PassIcon.tabProfileTrialUnselected
 
-        _ = PassIcon.tabAddWithBadge
         _ = PassIcon.tabMonitorActiveBreachesFoundSelected
         _ = PassIcon.tabMonitorActiveBreachesFoundUnselected
         _ = PassIcon.tabMonitorActiveNoBreachesSelected